Move chrome app to top level directory. Simplify Gruntfile.js, update readme, simplify ExtensionControllers
