+++ .gitignore
@@ -1,3 +1,15 @@
+node_modules
+temp
+.tmp
+dist
+.sass-cache
+app/bower_components
+test/bower_components
+package
+npm-debug.log
+
+# Generic stuff to ignore
+
 # Node.js with Express and Stlyus
 /.monitor
