+++ .jshintrc
