+++ app/manifest.json
@@ -1,7 +1,7 @@
 {
     "name": "__MSG_appName__",
     "description": "__MSG_appDescription__",
-    "version": "30",
+    "version": "31",
     "manifest_version": 2,
     "default_locale": "en",
     "permissions": [
