+++ app/scripts/ExtensionController.js
@@ -1,0 +1,9 @@
+/*global chrome*/
+
+/**
+ * Entry point for published Chrome App.
+ */
+
+import ChromeSerialBridge from './ChromeSerialBridge'
+
+window.chromeSerialBridge = new ChromeSerialBridge(chrome);
