+++ build-config.js
@@ -7,21 +7,8 @@
  */
 module.exports = {
   /**
-   * Should include properties:
-   * {
-   *   "AWSAccessKeyId": "MYACCESSKEYID",
-   *   "AWSSecretKey": "my+secret+key"
-   * }
-   */
-  "s3_keys_file": "~/.secrets/aws-keys.json",
-  /**
    * Generate following instructions from:
    * https://developer.chrome.com/extensions/packaging#creating
    */
   "chrome_signing_pem": "~/.secrets/chrome-app-test.pem",
-  /**
-   * Find the following in chrome://extensions with Developer mode checked
-   */
-  "chrome_deploy_extension_id": "hclkodmlbbhiknnciphlcnbkglgjakmd",
-  "chrome_dev_extension_id": "himpmjbkjeenflliphlaaeggkkanoglo"
 };
