--- chrome-app/.gitignore
@@ -1,7 +1,0 @@
-node_modules
-temp
-.tmp
-dist
-.sass-cache
-app/bower_components
-test/bower_components
