--- chrome-app/app/scripts/ExtensionController.js
@@ -1,11 +1,0 @@
-/*global chrome*/
-
-import ChromeSerialBridge from './ChromeSerialBridge'
-
-export default class ExtensionController {
-  constructor() {
-    this.chromeSerialBridge = new ChromeSerialBridge(chrome);
-  }
-}
-
-window.extensionController = new ExtensionController();
