--- chrome-app/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "chrome-fresh",
-  "version": "0.0.0",
-  "main": "dist/scripts/app.bundled.js",
-  "dependencies": {},
-  "devDependencies": {
-    "babel-preset-es2015": "^6.3.13",
-    "babelify": "^7.2.0",
-    "browser-serialport": "^2.0.3",
-    "browserify": "^13.0.0",
-    "grunt": "~0.4.1",
-    "grunt-browserify": "^4.0.1",
-    "grunt-chrome-manifest": "~0.2.0",
-    "grunt-concurrent": "~0.5.0",
-    "grunt-contrib-clean": "~0.5.0",
-    "grunt-contrib-compress": "~0.9.1",
-    "grunt-contrib-concat": "~0.3.0",
-    "grunt-contrib-connect": "~0.7.1",
-    "grunt-contrib-copy": "~0.5.0",
-    "grunt-contrib-cssmin": "~0.9.0",
-    "grunt-contrib-htmlmin": "~0.3.0",
-    "grunt-contrib-imagemin": "~0.7.1",
-    "grunt-contrib-jshint": "~0.9.2",
-    "grunt-contrib-uglify": "~0.4.0",
-    "grunt-contrib-watch": "~0.6.1",
-    "grunt-crx": "^1.0.3",
-    "grunt-exec": "^0.4.6",
-    "grunt-mocha": "~0.4.10",
-    "grunt-newer": "~0.6.0",
-    "grunt-open": "^0.2.3",
-    "grunt-svgmin": "~0.4.0",
-    "grunt-usemin": "~2.1.0",
-    "jshint-stylish": "~0.1.5",
-    "load-grunt-tasks": "~0.2.0",
-    "time-grunt": "~0.3.1"
-  },
-  "engines": {
-    "node": ">=0.8.0"
-  }
-}
